refactor: move main files to main folder
